requirements.txt
PyJWT>=2.8.0
openpyxl>=3.0.0
minio>=7.0.0
